+++ src/hooks.tsx
@@ -1,14 +1,16 @@
-import { ReactElement, useMemo } from "react";
+import { toJsxRuntime } from "hast-util-to-jsx-runtime";
+import { JSX, useMemo } from "react";
 import { Fragment, jsx, jsxs } from "react/jsx-runtime";
 import rehypeReact from "rehype-react";
 import remarkParse from "remark-parse";
 import remarkToRehype from "remark-rehype";
-import { unified } from "unified";
-import { UseRemarkOptions } from "./types.js";
-import { assertIsError } from "./utils.js";
+import { Processor, unified } from "unified";
 
-export function useRemark({ markdown, rehypePlugins = [], rehypeReactOptions, remarkParseOptions, remarkPlugins = [], remarkToRehypeOptions, onError = console.error }: UseRemarkOptions): ReactElement | null {
-  const processor = useMemo(
+import { UseRemarkOptions } from "./types.js";
+import { tryCatch } from "./utils.js";
+
+export function useRemark({ markdown, rehypePlugins = [], rehypeReactOptions, remarkParseOptions, remarkPlugins = [], remarkToRehypeOptions, components, onError = console.error }: UseRemarkOptions): JSX.Element | null {
+  const processor = useMemo<Processor>(
     () =>
       unified()
         .use(remarkParse, remarkParseOptions)
@@ -19,14 +21,13 @@ export function useRemark({ markdown, rehypePlugins = [], rehypeReactOptions, re
     []
   );
   const reactContent = useMemo(() => {
-    try {
-      const { result } = processor.processSync(markdown) as { result: ReactElement };
-      return result;
-    } catch (error) {
-      assertIsError(error);
-      onError(error);
-      return null;
-    }
+    const { success, data, error } = tryCatch(() => {
+      const file = processor.processSync(markdown);
+      return (components ? processor.runSync(processor.parse(file), file) : file.result) as any;
+    });
+    if (success) return components ? toJsxRuntime(data, { Fragment, components, ignoreInvalidStyle: true, jsx, jsxs, passKeys: true, passNode: true }) : data;
+    onError(error);
+    return null;
   }, [markdown]);
 
   return reactContent;